+++ tests/__init__.py
@@ -1,0 +1,3 @@
+import sys
+import tests.loguru as loguru_stub
+sys.modules['loguru'] = loguru_stub